Add aliasing to type names
When enabled with the "--aliases" command line option, the command
line tool will attempt to name the generated types using information
from the JSON key sets, instead of "generatedType3" and such.

This is off by default. If enabled, and --debug is enabled, then
the output includes the internal generated name as a comment in the
source code to assist in debugging.

Also added --version to the command line to print the version number
and Go build system used.