Add --suffix option which specifies string added to type names
The default is to add "Type" after each type name, to distinguish
it from member names in the JSON objects. If this result in a name
collision, you can use the --suffix option to specify an alternate
string suffix to apply to type names.

This fix also removes an unwanted debugging message, and updates
the README file.