Detect sparse field types, automatically add omitempty to Go
If a field is used in an array in the JSON, and all of its fields
match an existing type, then it is considered a match. If the value
does not have all the memebers of the type, it is still a match but
the field(s) that are missing from the test type are marked as being
able to be omitted. This causes the output to automatically generate
"omitempty" json tags.  --omit can still be used to force this for
all field members.

For swift, any fields not present in all the objects in the source
json are marked as Swift optionals. For Swift, turned off pretty-print
by default so columns are not forced into alignment in the declarations,
to model how XCode looks. You can use --pretty to turn it back on.

Additionally, alias usage is the default, you must now specify
--no-alias if you want to disable this feature.